+++ examples/multi_page_0.py
@@ -4,46 +4,46 @@ from gladius import Gladius, Component, EventRequest
 from gladius.daisyui import DaisyUI
 
 g = Gladius()
-dui = DaisyUI(g)
+d = DaisyUI(g)
 
 def my_navbar() -> Component:
-    navbar = dui.Navbar(hx_boost='true')
-    navbar.add(a := dui.Link(href='/').add(dui.Text('Home')))
-    navbar.add(a := dui.Link(href='/dashboard').add(dui.Text('Dashboard')))
-    navbar.add(a := dui.Link(href='/signin').add(dui.Text('Sign In')))
-    navbar.add(a := dui.Link(href='/signup').add(dui.Text('Sign Up')))
+    navbar = d.Navbar(hx_boost='true')
+    navbar.add(a := d.Link(href='/').add(d.Text('Home')))
+    navbar.add(a := d.Link(href='/dashboard').add(d.Text('Dashboard')))
+    navbar.add(a := d.Link(href='/signin').add(d.Text('Sign In')))
+    navbar.add(a := d.Link(href='/signup').add(d.Text('Sign Up')))
     return navbar
 
 def my_join() -> Component:
-    join = dui.Join(hx_boost='true')
-    join.add(a := dui.Link(class_='btn join-item', href='/').add(dui.Text('Home')))
-    join.add(a := dui.Link(class_='btn join-item', href='/dashboard').add(dui.Text('Dashboard')))
-    join.add(a := dui.Link(class_='btn join-item', href='/signin').add(dui.Text('Sign In')))
-    join.add(a := dui.Link(class_='btn join-item', href='/signup').add(dui.Text('Sign Up')))
+    join = d.Join(hx_boost='true')
+    join.add(a := d.Link(class_='btn join-item', href='/').add(d.Text('Home')))
+    join.add(a := d.Link(class_='btn join-item', href='/dashboard').add(d.Text('Dashboard')))
+    join.add(a := d.Link(class_='btn join-item', href='/signin').add(d.Text('Sign In')))
+    join.add(a := d.Link(class_='btn join-item', href='/signup').add(d.Text('Sign Up')))
     return join
 
 def root_page() -> Component:
     # root/index
-    page = dui.Page(title='Multi Page 0')
+    page = d.Page(title='Multi Page 0')
 
     # navbar
     page.add(navbar := my_navbar())
 
     # join buttons
-    page.add(flex := dui.Flex())
+    page.add(flex := d.Flex())
     flex.add(join := my_join())
     return page
 
 def dashboard_page() -> Component:
-    page = dui.Page(title='Dashboard')
+    page = d.Page(title='Dashboard')
     page.add(navbar := my_navbar())
     
     # vflex - Hello World
-    page.add(vflex := dui.VFlex())
-    vflex.add(card := dui.Card().add(dui.Text('Hello')))
-    vflex.add(flex := dui.Flex())
-    flex.add(card := dui.Card().add(dui.Text('World')))
-    flex.add(card := dui.Card().add(dui.Text('!')))
+    page.add(vflex := d.VFlex())
+    vflex.add(card := d.Card().add(d.Text('Hello')))
+    vflex.add(flex := d.Flex())
+    flex.add(card := d.Card().add(d.Text('World')))
+    flex.add(card := d.Card().add(d.Text('!')))
     
     async def add_button_click(button, req: EventRequest):
         row = [randint(0, 10) for i in range(3)]
@@ -57,29 +57,29 @@ def dashboard_page() -> Component:
         table_1.rows = [*table_1.rows[:-1]]
     
     # join button
-    vflex.add(join := dui.Join())
-    join.add(add_button := dui.Button(onclick=add_button_click).add(dui.Text('Add')))
-    join.add(edit_button := dui.Button(class_='btn btn-secondary').add(dui.Text('Edit')))
-    join.add(remove_button := dui.Button(onclick=remove_button_click, class_='btn btn-primary').add(dui.Text('Remove')))
+    vflex.add(join := d.Join())
+    join.add(add_button := d.Button(onclick=add_button_click).add(d.Text('Add')))
+    join.add(edit_button := d.Button(class_='btn btn-secondary').add(d.Text('Edit')))
+    join.add(remove_button := d.Button(onclick=remove_button_click, class_='btn btn-primary').add(d.Text('Remove')))
 
     # table_0
     header = ['A', 'B', 'C']
     rows = [[j for j in range(i, i + 3)] for i in range(5)]
-    vflex.add(table_0 := dui.Table(header=header, rows=rows))
+    vflex.add(table_0 := d.Table(header=header, rows=rows))
     
     # table_1
-    header = ['A', 'B', 'C']
+    header = ['D', 'E', 'F']
     rows = [[j for j in range(i, i + 3)] for i in range(100, 105)]
-    vflex.add(table_1 := dui.Table(header=header, rows=rows))
+    vflex.add(table_1 := d.Table(header=header, rows=rows))
     return page
 
 def sign_in_page() -> Component:
-    page = dui.Page(title='Sign In')
+    page = d.Page(title='Sign In')
     page.add(navbar := my_navbar())
     return page
 
 def sign_up_page() -> Component:
-    page = dui.Page(title='Sign In')
+    page = d.Page(title='Sign In')
     page.add(navbar := my_navbar())
     return page